requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
jinja2==3.1.2
pyyaml==6.0.1
httpx==0.26.0
aiofiles==23.2.1
python-dotenv==1.0.0
